+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added a Cloudformation stack with template url and policy url stored in an s3 bucket
 - removed aws_cloudformation_type resource from the module
 - Commented out deployment_targets since auto_deployment feature is required for SERVICE_MANAGED permission model.The feature will be added in subsequent releases.
+- modified policy_body and disable_rollback conditions
+- added a random string to example bucket name
 
 ## [1.1.2] - 2023-09-05
   - fix: CKV_AWS_109 "Ensure IAM policies does not allow permissions management / resource exposure without constraints"